+++ containertemplates/tests/factories.py
@@ -1,22 +1,8 @@
 """Factories for container tests."""
 
 import factory
-from projectroles.constants import SODAR_CONSTANTS
-from projectroles.models import Project
 
 from containertemplates.models import ContainerTemplate
-
-
-class ProjectFactory(factory.django.DjangoModelFactory):
-    """Factory for creating ``projectroles`` ``Project`` objects."""
-
-    class Meta:
-        model = Project
-
-    title = factory.Sequence(lambda n: "Project %03d" % n)
-    type = SODAR_CONSTANTS["PROJECT_TYPE_PROJECT"]
-    parent = None
-    description = factory.Sequence(lambda n: "This is project %03d" % n)
 
 
 class ContainerTemplateFactory(factory.django.DjangoModelFactory):
@@ -29,7 +15,6 @@ class ContainerTemplateFactory(factory.django.DjangoModelFactory):
     description = "Some description"
     repository = factory.Sequence(lambda n: f"repository{n}")
     tag = "latest"
-    project = factory.SubFactory(ProjectFactory)
     container_port = 80
     container_path = ""
     heartbeat_url = ""